Move repository package to internal/repository
